+++ mywords-flutter/lib/pages/statistic_chart.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mywords/libso/handler_for_native.dart'
     if (dart.library.html) 'package:mywords/libso/handler_for_web.dart';
+import 'package:mywords/util/get_scaffold.dart';
 
 import 'package:mywords/widgets/line_chart.dart';
 
@@ -88,7 +89,7 @@ class _State extends State<StatisticChart> with SingleTickerProviderStateMixin {
     );
     return DefaultTabController(
       length: myTabs.length,
-      child: Scaffold(
+      child: getScaffold(context,
         appBar: appBar,
         body: TabBarView(children: tableWidgets),
       ),
