+++ mywords-go/.gitignore
@@ -25,3 +25,4 @@ vendor/
 
 *.profile
 _*.*
+cmd/flutter/web/*
