+++ mywords-go/cmd/flutter/file_handler.go
@@ -1,0 +1,17 @@
+package main
+
+import (
+	"embed"
+	"io/fs"
+	"path/filepath"
+)
+
+type webEmbedHandler struct {
+	webEmbed embed.FS
+}
+
+func (f webEmbedHandler) Open(name string) (fs.File, error) {
+	// 在windows系统下必须用toSlash 封装一下路径，否则，web\index.html!=web/index.html
+	name = filepath.ToSlash(filepath.Join("web", name))
+	return f.webEmbed.Open(name)
+}
